Moved extractSnippetValues to a separate file
